+++ src/Language/PHP/Pretty.hs
@@ -269,7 +269,8 @@ prettyPropertyDecl (PropertyDecl annotation attrs modif mType items hooks) = pre
     else " {" <> line <> indent 4 (vsep (map prettyHook hooks)) <> line <> "}"
 
 prettyHook :: HasLeadingTrivia a => PropertyHook a -> Doc ann
-prettyHook (PropertyHook annotation hookT mParam body) = prettyLeadingTrivia annotation $
+prettyHook (PropertyHook annotation isFinal hookT mParam body) = prettyLeadingTrivia annotation $
+  (if isFinal then "final " else "") <>
   (case hookT of HookGet -> "get"; HookSet -> "set") <>
   maybe mempty (\(var, mTyp) -> "(" <> maybe mempty (\t -> prettyType t <> " ") mTyp <> prettyVarName var <> ")") mParam <>
   case body of